Finishing Test Level
The test level is almost complete, with the interior being all sealed and ceilings being added mostly everywhere. All that remains at the moment (for the geometry) is to put in more decorations, lights and to finish up the center ceiling. Soon enough the player may be free to walk around on the ceiling before finally hopping in through the skylight to start having fun.