Finished Test Level
The test level now has boundaries and proper lighting, along with a well defined entry point from the ceiling into the level itself. All that remains now is to replace various props with their respective models, as well as add pickups and the enemies, along with a level clear interactable.