More Props in Level 2
Added more props and general structure in the lobby level, helping flesh out the environment more by making a kitchen, a small storage area and adding some additional paintings at the top.

All that remains at the moment are fleshing out the lone corner on the first floor as well as the second floor overall.